refactor: imporve the structures
