+++ src/components/Todo.jsx
@@ -1,5 +1,6 @@
 import React, { useMemo } from 'react';
 import { LazyImage } from '.';
+import getNames from '../utils/getName';
 import './Todo.css';
 
 export const Todo = ({ todo, onChange, config, isCompleted }) => {
@@ -27,53 +28,12 @@ export const Todo = ({ todo, onChange, config, isCompleted }) => {
     }
   }, [config.sorted]);
 
-  var formattedNames = [];
-
-  const getNames = (names) => {
-    if (names === Object(names)) {
-      // Object here
-      if (names.name != undefined && names.name != null) {
-        if (Object.keys(names.name).length > !2) {
-          for (const [key, value] of Object.entries(names.name)) {
-            formattedNames += value;
-          }
-        } else {
-          for (const [key, value] of Object.entries(names.name)) {
-            formattedNames += key + value;
-          }
-        }
-        return formattedNames;
-      } else {
-        return 'No name';
-      }
-    } else if (Array.isArray(names)) {
-      if (names.name != undefined && names.name != null) {
-        if (names.name.length > !2) {
-          for (const [key, value] of Object.entries(names.name)) {
-            formattedNames += value;
-            return formattedNames;
-          }
-        } else {
-          for (const [key, value] of Object.entries(names.name)) {
-            formattedNames += key + value;
-            return formattedNames;
-          }
-        }
-      } else {
-        return 'No name';
-      }
-    }
-  };
-
-  const userName = useMemo(() => getNames(todo), []);
+  const userName = useMemo(() => getNames(todo), [todo]);
 
   const getLocation = ({ location: { street, ...rest } }) => {
     const myStreetName = street.name;
     const myStreetNumber = street.number;
     const myPostcode = rest.postcode;
-    const myCity = rest.city;
-    const myState = rest.state;
-    const myCountry = rest.country;
 
     return (
       <>
@@ -85,7 +45,7 @@ export const Todo = ({ todo, onChange, config, isCompleted }) => {
     );
   };
 
-  const userLocation = useMemo(() => getLocation(todo), []);
+  const userLocation = useMemo(() => getLocation(todo), [todo]);
 
   const getPicture = () => {
     return <LazyImage src={todo.picture.large} alt={todo.picture.large} />;
